Chrome changes to support WebKit API changes.
Also, I removed the GetWebFrame method on WebDataSource since it is not
actually needed.

Removed some dead-code from webframe_impl.cc.

Removed some bogus null-checking of WebCore::Frame::loader().

R=dglazkov

Review URL: http://codereview.chromium.org/67169

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13844 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/renderer/renderer_webkitclient_impl.cc
@@ -24,7 +24,7 @@ using WebKit::WebURL;
 
 WebKit::WebClipboard* RendererWebKitClientImpl::clipboard() {
   return &clipboard_;
-} 
+}
 
 WebKit::WebMimeRegistry* RendererWebKitClientImpl::mimeRegistry() {
   return &mime_registry_;
@@ -38,13 +38,14 @@ WebKit::WebSandboxSupport* RendererWebKitClientImpl::sandboxSupport() {
 #endif
 }
 
-uint64_t RendererWebKitClientImpl::visitedLinkHash(const char* canonical_url,
-                                                   size_t length) {
+unsigned long long RendererWebKitClientImpl::visitedLinkHash(
+    const char* canonical_url,
+    size_t length) {
   return RenderThread::current()->visited_link_slave()->ComputeURLFingerprint(
       canonical_url, length);
 }
 
-bool RendererWebKitClientImpl::isLinkVisited(uint64_t link_hash) {
+bool RendererWebKitClientImpl::isLinkVisited(unsigned long long link_hash) {
   return RenderThread::current()->visited_link_slave()->IsVisited(link_hash);
 }
